Tenancy based on Organization:: tests passed

+++ app/Http/Controllers/Api/OrganizationController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use App\Models\Organization;
+use App\Permissions\OrganizationPermissions;
+use App\Roles\OrganizationRoles;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Gate;
 
@@ -11,6 +13,8 @@ class OrganizationController extends Controller
 {
     public function index()
     {
+        Gate::authorize('viewAny', Organization::class);
+
         return response()->json([
             'data' => Organization::all()
         ]);
@@ -18,6 +22,16 @@ class OrganizationController extends Controller
 
     public function store(Request $request)
     {
+        $user = auth()->user();
+        // dd([
+        //     $user->organizations,
+        //     $user->getAllRoles(),
+        //     $user->getAllPermissions(),
+        //     $user->hasPermission(OrganizationPermissions::CREATE_ORGANIZATION),
+        //     $user->hasRole(OrganizationRoles::SUPER_ADMIN)
+        // ]);
+        Gate::authorize('create', Organization::class);
+
         $validated = $request->validate([
             'name' => 'required|string|max:255|unique:organizations',
             'description' => 'nullable|string'
@@ -32,6 +46,8 @@ class OrganizationController extends Controller
 
     public function show(Organization $organization)
     {
+        Gate::authorize('view', [Organization::class, $organization]);
+
         return response()->json([
             'data' => $organization
         ]);
@@ -39,6 +55,16 @@ class OrganizationController extends Controller
 
     public function update(Request $request, Organization $organization)
     {
+        $user = auth()->user();
+        // dd([
+        //     $user->organizations,
+        //     $user->getAllRoles(),
+        //     $user->getAllPermissions(),
+        //     $user->hasPermission(OrganizationPermissions::CREATE_ORGANIZATION),
+        //     $user->hasRole(OrganizationRoles::SUPER_ADMIN)
+        // ]);
+        Gate::authorize('update', [Organization::class, $organization]);
+
         $validated = $request->validate([
             'name' => 'sometimes|required|string|max:255|unique:organizations,name,' . $organization->id,
             'description' => 'nullable|string'
@@ -53,6 +79,8 @@ class OrganizationController extends Controller
 
     public function destroy(Organization $organization)
     {
+        Gate::authorize('delete', [Organization::class, $organization]);
+
         $organization->delete();
 
         return response()->json(null, 204);
